+++ .remarkrc.js
@@ -1,6 +1,17 @@
+const unified = require("unified")
+
 module.exports = {
   plugins: [
     ["remark-frontmatter", "yaml"],
+    [
+      "remark-retext",
+      unified()
+        .use(require("retext-english"))
+        .use(require("retext-diacritics"))
+        .use(require("retext-indefinite-article"))
+        .use(require("retext-redundant-acronyms"))
+        .use(require("retext-sentence-spacing")),
+    ],
     "remark-preset-lint-recommended",
     "remark-preset-lint-markdown-style-guide",
 
@@ -13,7 +24,6 @@ module.exports = {
     // We would like these rules to be enabled, but they require significant content changes
     // and need additional PRs to be implemented
     ["remark-lint-emphasis-marker", false],
-    ["remark-lint-heading-increment", false],
     ["remark-lint-no-heading-punctuation", false],
     ["remark-lint-list-item-spacing", false],
 
@@ -35,5 +45,6 @@ module.exports = {
     ["remark-lint-no-shortcut-reference-link", false],
     // We use brackets in a lot of places as argument lists and do not want to escape them.
     ["remark-lint-no-undefined-references", false],
+    ["remark-lint-first-heading-level", 2]
   ],
 }