+++ packages/gatsby-telemetry/src/telemetry.js
@@ -196,11 +196,7 @@ module.exports = class AnalyticsTracker {
   }
 
   getDbEngine() {
-    // if (process.env.GATSBY_DB_NODES === `loki`) {
-    //   return `loki`
-    // } else {
     return `redux`
-    // }
   }
 
   getMachineId() {