+++ packages/gatsby-remark-autolink-headers/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@2.3.4...gatsby-remark-autolink-headers@2.3.5) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+## [2.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@2.3.3...gatsby-remark-autolink-headers@2.3.4) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+## [2.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@2.3.2...gatsby-remark-autolink-headers@2.3.3) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+## [2.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@2.3.1...gatsby-remark-autolink-headers@2.3.2) (2020-05-13)
+
+### Bug Fixes
+
+- **gatsby-remark-autolink-headers:** include scroll-margin in getTargetOffset ([#24018](https://github.com/gatsbyjs/gatsby/issues/24018)) ([3696858](https://github.com/gatsbyjs/gatsby/commit/3696858))
+
 ## [2.3.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-autolink-headers@2.3.0...gatsby-remark-autolink-headers@2.3.1) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-remark-autolink-headers