+++ packages/gatsby-remark-prismjs/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-prismjs@3.5.3...gatsby-remark-prismjs@3.5.4) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-remark-prismjs
+
+## [3.5.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-prismjs@3.5.2...gatsby-remark-prismjs@3.5.3) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-remark-prismjs
+
+## [3.5.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-prismjs@3.5.1...gatsby-remark-prismjs@3.5.2) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-remark-prismjs
+
 ## [3.5.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-prismjs@3.5.0...gatsby-remark-prismjs@3.5.1) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-remark-prismjs