+++ packages/gatsby-source-npm-package-search/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.5...gatsby-source-npm-package-search@2.3.6) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+## [2.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.4...gatsby-source-npm-package-search@2.3.5) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+## [2.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.3...gatsby-source-npm-package-search@2.3.4) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+## [2.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.2...gatsby-source-npm-package-search@2.3.3) (2020-05-12)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+## [2.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.1...gatsby-source-npm-package-search@2.3.2) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
 ## [2.3.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.3.0...gatsby-source-npm-package-search@2.3.1) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-source-npm-package-search