+++ packages/gatsby-theme/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme@0.1.1...gatsby-theme@0.1.2) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-theme
+
 ## [0.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme@0.1.0...gatsby-theme@0.1.1) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-theme